IGNITE-24846 Move NOOP to GridTestUtils
