+++ modules/core/src/test/java/org/apache/ignite/spi/communication/tcp/TcpCommunicationSpiHalfOpenedConnectionTest.java
@@ -35,7 +35,7 @@ import org.apache.ignite.testframework.GridTestUtils;
 import org.apache.ignite.testframework.junits.common.GridCommonAbstractTest;
 import org.junit.Test;
 
-import static org.apache.ignite.internal.util.lang.ClusterNodeFunc.noop;
+import static org.apache.ignite.testframework.GridTestUtils.noop;
 
 /**
  * Tests case when connection is closed only for one side, when other is not notified.
